backend/app/requirements.txt

# ============================================================================
# Production Dependencies
# ============================================================================

# Web Framework & Server
fastapi
uvicorn[standard]
requests 

# Database & ORM
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic

# Configuration & Environment
python-dotenv
pydantic
pydantic-settings

# Authentication & Security
python-jose[cryptography]
passlib[argon2]
argon2-cffi
cryptography
google-auth